requirements.txt
pre-commit==2.6.0
django==3.1
Pillow==7.2.0
numpy==1.15.4
